Added --ignorecase option to choose case-insensitive searching.
The default search is now case sensitive.

There is also a change in behavior in that previously note titles would
be displayed if they matched the search pattern regardless of whether
lines in the note matched as well.  Now note titles are ignored when
searching.

(Technically the evernote searching still searches against note titles,
but they are now filtered out unless the note has matching lines as
well.)